refac(rs): move point implementaitons to `short_weierstrass`
